Support iron 0.5.x
From iron 0.5.x, iron::Request includes a private field, so it is not
feasible to create iron::Request directly. So, we need to make
hyper::client::Request to make iron::Request.

hyper::client::Request also contains a private field. The public
interface to make it is to parse a http request.

So, in request.rs, we generate http request string, and pass it to
hyper::client::Request::new.